I am updating the version of CLUBB used in E3SM-CLUBB-SILHS to be based on CLUBB 36e1290 from the CLUBB master branch.  A modified version for backward compatibility to r8220, plus E3SM customizations, is saved the the CLUBB branch clubb_E3SM_silhs.
The 36e1290 revision was chosen because it is the same revision
that is used to base E3SM-CLUBB for the next version of the
master branch of E3SM.  This is the version found in
singhbalwinder/atm/clubbv2.

This version of the code will not compile until corresponding changes
are made to clubb_intr.F90.