Explicitly specify srun args for running on cori-knl
This is a tentative fix to get the branch to run on cori-knl.

The following changes are made in cime/cime_config/acme/machines/config_machines.xml
in the section of cori-knl

from
      <arg name="num_tasks" > -n $TOTALPES</arg>
      <arg name="thread_count">-c $SHELL{echo 272/`./xmlquery --value MAX_MPITASKS_PER_NODE`|bc}</arg>
      <arg name="binding"> $SHELL{if [ 68 -ge `./xmlquery --value MAX_MPITASKS_PER_NODE` ]; then echo "--cpu_bind=cores"; else echo "--cpu_bind=threads";fi;} </arg>
to

      <arg name="num_tasks" > -n 2560</arg>
      <arg name="thread_count">-c 8</arg>
      <arg name="binding">--cpu_bind=cores</arg>

The changes make the run on cori-knl not flexible and can only use a fixed 2560 tasks on 80 nodes.

[BFB]